Better?
In Windows it didn't appear to matter, but Linux did give some ugly output.
Hopefully that is no longer the case.

NOTE! If this produces better output, I am NOT going to recompile and re-release all Linux tools because of this. Any Linux tool that still needed to be done anyway, or was about to be updated anyway will benefit. That's all.